Back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = ""
authors = [
    {name = "linhphung211",email = "lphungtuan2004@gmail.com"}
]

requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "markdown (>=3.10.2,<4.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-jazzmin (>=3.0.4,<4.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"